requirements.txt
hydra-core>=1.2.0
wandb>=0.13
numpy
torch>=1.10
transformers>=4.22.0
tokenizers
swig>=4.1.0
box2d-py==2.3.8
requests
Flask
scikit-learn
matplotlib
langchain==0.0.225
typing_extensions==4.5
openai
plotly
